pyproject.toml
[tool.poetry]
name = "datastore"
version = "0.0.24"
description = "Config driven Data Ingestion Framework"
authors = ["Krijn van der Burg <krijnvdburg@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pyspark = "^3.5.0"
pyarrow = "^14.0.2"
pyyaml = "^6.0.1"
jsonschema = "^4.21.1"

[tool.poetry.group.test.dependencies]
pytest = "^8.1.1"
pandas = "^2.2.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
